Add support direct passing raw content type in attachment and embeds
Now users can pass any allowed content type in attachments and embeds (images and thumbnails)
directly as bytes or file-like objects. It will be converted to
relevant data and passed to the API.
It will simplify the process of sending attachment and images to embeds.

Also I allowed to pass single embed instead of list of embeds in message
data. It will be converted to list of embeds automatically.